+++ tests/EasyDBTest.php
@@ -4,7 +4,6 @@ namespace ParagonIE\EasyDB\Tests;
 use ParagonIE\EasyDB\Factory;
 use PDO;
 use PHPUnit_Framework_TestCase;
-use ReflectionClass;
 
 /**
  * Class EasyDBTest
@@ -73,37 +72,6 @@ abstract class EasyDBTest
                 };
             },
             $this->GoodFactoryCreateArgumentProvider()
-        );
-    }
-
-    /**
-    * EasyDB data provider
-    * Returns an array of callables that return instances of EasyDB
-    * @return array
-    * @see EasyDBTest::GoodFactoryCreateArgumentProvider()
-    */
-    public function GoodFactoryCreateArgument2EasyDBWithPDOAttributeProvider()
-    {
-        $ref = new ReflectionClass(PDO::class);
-        $attrs = array_filter(
-            $ref->getConstants(),
-            function ($attrName) {
-                return (strpos($attrName, 'ATTR_') === 0);
-            },
-            ARRAY_FILTER_USE_KEY
-        );
-        return array_reduce(
-            $this->GoodFactoryCreateArgument2EasyDBProvider(),
-            function (array $was, callable $cb) use ($attrs) {
-                foreach ($attrs as $attr) {
-                    $was[] = [
-                        $cb,
-                        $attr
-                    ];
-                }
-                return $was;
-            },
-            []
         );
     }
 }
