+++ tests/GetAttributeTest.php
@@ -5,11 +5,43 @@ use ParagonIE\EasyDB\EasyDB;
 use ParagonIE\EasyDB\Factory;
 use PDO;
 use PDOException;
+use ReflectionClass;
 
 class GetAttributeTest
     extends
         EasyDBTest
 {
+
+    /**
+    * EasyDB data provider
+    * Returns an array of callables that return instances of EasyDB
+    * @return array
+    * @see EasyDBTest::GoodFactoryCreateArgument2EasyDBProvider()
+    */
+    public function GoodFactoryCreateArgument2EasyDBWithPDOAttributeProvider()
+    {
+        $ref = new ReflectionClass(PDO::class);
+        $attrs = array_filter(
+            $ref->getConstants(),
+            function ($attrName) {
+                return (strpos($attrName, 'ATTR_') === 0);
+            },
+            ARRAY_FILTER_USE_KEY
+        );
+        return array_reduce(
+            $this->GoodFactoryCreateArgument2EasyDBProvider(),
+            function (array $was, callable $cb) use ($attrs) {
+                foreach ($attrs as $attr) {
+                    $was[] = [
+                        $cb,
+                        $attr
+                    ];
+                }
+                return $was;
+            },
+            []
+        );
+    }
 
     /**
     * @dataProvider GoodFactoryCreateArgument2EasyDBWithPDOAttributeProvider
